+++ TodoMateTests/Mock/MockFetchAuthenticatedUserUseCase.swift
@@ -1,0 +1,15 @@
+//
+//  MockFetchAuthenticatedUserUseCase.swift
+//  TodoMate
+//
+//  Created by hs on 3/10/25.
+//
+
+@testable import TodoMate
+
+struct MockFetchAuthenticatedUserUseCase: FetchAuthenticatedUserUseCaseType {
+    private let excuteReturn: AuthenticatedUser?
+    init(excuteReturn: AuthenticatedUser? = nil) { self.excuteReturn = excuteReturn }
+
+    func execute() -> AuthenticatedUser? { excuteReturn }
+}
